DeliveryServer: rename testRunnerAdapter to DeliveryContainer
