+++ views/js/controller/runtime/service/fullScreen.js
