+++ views/templates/DeliveryServer/container/service/template.tpl
